use a minimum bounding box algorithm to recover the frame of a table
reuses code for continuous walking block segmentation.
in particular gives more reliable standing frame
still needs some improvement but works better than previous approach

edgePoint is no longer required for segmentTableEdge, but haven't change it
